@@ -2,35 +2,35 @@
   "name": "trainboard",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "dependencies": {
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.11.16",
-    "@mui/styles": "^5.11.16",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^6.3.0",
+    "@mui/material": "^6.3.0",
+    "@mui/styles": "^6.3.0",
     "date-fns": "^2.29.3",
     "date-fns-tz": "^2.0.0",
-    "google-map-react": "^2.2.0",
     "gtfs-realtime-bindings": "^1.1.1",
-    "prettier": "^2.8.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "prettier": "^3.4.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-router-dom": "^6.10.0",
     "use-debounce": "^9.0.4"
   },
   "devDependencies": {
-    "@types/google-map-react": "^2.1.7",
     "@types/node": "^18.15.11",
-    "@types/react": "^17.0.56",
-    "@types/react-dom": "^17.0.19",
-    "@vitejs/plugin-react": "^4.2.1",
+    "@types/react": "^19.0.2",
+    "@types/react-dom": "^19.0.2",
+    "@vitejs/plugin-react": "^4.3.4",
     "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-unused-imports": "^2.0.0",
-    "husky": "^8.0.0",
+    "husky": "^9.1.7",
+    "lint-staged": "^15.2.11",
     "sass": "^1.61.0",
-    "typescript": "^4.9.5",
-    "vite": "^5.1.6",
-    "vite-plugin-svgr": "^4.2.0"
+    "typescript": "^5.7.2",
+    "vite": "^6.0.5",
+    "vite-plugin-svgr": "^4.3.0"
   },
   "scripts": {
     "start": "vite",
@@ -41,7 +41,7 @@
     "lint:fix": "npm run lint:eslint -- --fix",
     "typecheck": "tsc --noEmit",
     "tunnel": "ngrok http 3000",
-    "prepare": "husky install"
+    "prepare": "husky"
   },
   "browserslist": {
     "production": [
